pyproject.toml
[project]
name = "agent-service-toolkit"
version = "0.1.0"
description = "Full toolkit for running an AI agent service built with LangGraph, FastAPI and Streamlit"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"

dependencies = [
    "duckduckgo-search>=8.0.1",
    "fastapi>=0.115.12",
    "grpcio>=1.71.0",
    "httpx>=0.27.2",
    "jiter>=0.8.2",
    "langchain-core>=0.3.59",
    "langchain-community>=0.3.23",
    "langchain-anthropic>=0.3.13",
    "langchain-aws>=0.2.23",
    "langchain-chroma>=0.2.3",
    "langchain-google-genai>=2.0.11",
    "langchain-google-vertexai>=2.0.7",
    "langchain-groq>=0.2.5",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.16",
    "langgraph>=0.3.34",
    "langgraph-checkpoint-mongodb>=0.1.3",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "langgraph-supervisor>=0.0.21",
    "langsmith>=0.1.147",
    "numexpr>=2.10.2",
    "numpy>=1.26.4; python_version <= '3.12'",
    "numpy>=2.2.3; python_version >= '3.13'",
    "onnxruntime>=1.22.0",
    "pandas>=2.2.3",
    "psycopg[binary,pool]>=3.2.7",
    "pyarrow>=20.0.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.6.1",
    "pyowm>=3.3.0",
    "python-dotenv>=1.0.1",
    "setuptools>=75.6.0",
    "streamlit>=1.40.2",
    "tiktoken>=0.9.0",
    "uvicorn>=0.32.1",
    "pypdf>=5.3.1",
    "docx2txt>=0.9",
    "langchain-mcp-adapters>=0.1.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.9",
    "mypy>=1.15.0",
]

# Group for the minimal dependencies to run just the client and Streamlit app.
# These are also installed in the default dependencies.
# To install run: `uv sync --frozen --only-group client`
client = [
    "httpx>=0.27.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"

[tool.mypy]
plugins = "pydantic.mypy"
exclude = "src/streamlit_app.py"

[[tool.mypy.overrides]]
module = ["numexpr.*"]
follow_untyped_imports = true
